Extract ingress topology read registrars (#1730)
