+++ control_plane/http_app.py
@@ -46,6 +46,11 @@ from control_plane import secrets as control_plane_secrets
 from control_plane import service_status as control_plane_service_status
 from control_plane import tracked_target_logs as control_plane_tracked_target_logs
 from control_plane import live_target_runtime as control_plane_live_target_runtime
+from control_plane.http_routes import (
+    ReadRouteDependencies,
+    register_ingress_read_routes,
+    register_topology_read_routes,
+)
 from control_plane.agent_context_service import (
     AgentContextPayload,
     agent_context_action_allowed,
@@ -536,7 +541,6 @@ from control_plane.contracts.protected_artifacts import (
 )
 from control_plane.contracts.private_health_endpoint_record import PrivateHealthEndpointRecord
 from control_plane.contracts.route_binding_record import (
-    EnvironmentRouteBindingReadModel,
     EnvironmentRouteBindingRecord,
     redacted_route_binding_record,
 )
@@ -1591,104 +1595,6 @@ class TrackedTargetLogsResponse(BaseModel):
     request: TrackedTargetLogRequestResponse
     logs: TrackedTargetLogLinesResponse
     deployment: TrackedTargetLogDeploymentResponse | None = None
-
-
-class EdgeEndpointRecordResponse(BaseModel):
-    model_config = ConfigDict(extra="forbid")
-
-    status: Literal["ok"] = "ok"
-    trace_id: str
-    record: EdgeEndpointRecord
-
-
-class EdgeEndpointRecordsResponse(BaseModel):
-    model_config = ConfigDict(extra="forbid")
-
-    status: Literal["ok"] = "ok"
-    trace_id: str
-    limit: int
-    count: int
-    records: tuple[EdgeEndpointRecord, ...]
-
-
-class PrivateHealthEndpointRecordResponse(BaseModel):
-    model_config = ConfigDict(extra="forbid")
-
-    status: Literal["ok"] = "ok"
-    trace_id: str
-    record: PrivateHealthEndpointRecord
-
-
-class PrivateHealthEndpointRecordsResponse(BaseModel):
-    model_config = ConfigDict(extra="forbid")
-
-    status: Literal["ok"] = "ok"
-    trace_id: str
-    product: str
-    context: str
-    instance: str
-    limit: int
-    count: int
-    records: tuple[PrivateHealthEndpointRecord, ...]
-
-
-class RouteBindingRecordResponse(BaseModel):
-    model_config = ConfigDict(extra="forbid")
-
-    status: Literal["ok"] = "ok"
-    trace_id: str
-    record: EnvironmentRouteBindingReadModel
-
-
-class RouteBindingRecordsResponse(BaseModel):
-    model_config = ConfigDict(extra="forbid")
-
-    status: Literal["ok"] = "ok"
-    trace_id: str
-    product: str
-    context: str
-    instance: str
-    limit: int
-    count: int
-    records: tuple[EnvironmentRouteBindingReadModel, ...]
-
-
-class IngressCanaryRouteRecordResponse(BaseModel):
-    model_config = ConfigDict(extra="forbid")
-
-    status: Literal["ok"] = "ok"
-    trace_id: str
-    record: IngressCanaryRouteRecord
-
-
-class IngressCanaryRouteRecordsResponse(BaseModel):
-    model_config = ConfigDict(extra="forbid")
-
-    status: Literal["ok"] = "ok"
-    trace_id: str
-    limit: int
-    count: int
-    records: tuple[IngressCanaryRouteRecord, ...]
-
-
-class IngressRouteAuditRecordResponse(BaseModel):
-    model_config = ConfigDict(extra="forbid")
-
-    status: Literal["ok"] = "ok"
-    trace_id: str
-    record: IngressRouteAuditRecord
-
-
-class IngressRouteAuditRecordsResponse(BaseModel):
-    model_config = ConfigDict(extra="forbid")
-
-    status: Literal["ok"] = "ok"
-    trace_id: str
-    product: str
-    context: str
-    limit: int
-    count: int
-    records: tuple[IngressRouteAuditRecord, ...]
 
 
 class EveryCodeWorkRequestRecordsResponse(BaseModel):
@@ -3092,79 +2998,6 @@ class _RecentOperationsReadStore(Protocol):
     ) -> tuple[PreviewRecord, ...]: ...
 
 
-class _EdgeEndpointReadStore(Protocol):
-    def read_edge_endpoint_record(self, endpoint_key: str) -> EdgeEndpointRecord: ...
-
-    def list_edge_endpoint_records(
-        self,
-        *,
-        provider: str = "",
-        status: str = "",
-        limit: int | None = None,
-    ) -> tuple[EdgeEndpointRecord, ...]: ...
-
-
-class _PrivateHealthEndpointReadStore(Protocol):
-    def read_private_health_endpoint_record(
-        self, endpoint_key: str
-    ) -> PrivateHealthEndpointRecord: ...
-
-    def list_private_health_endpoint_records(
-        self,
-        *,
-        product: str = "",
-        context_name: str = "",
-        instance_name: str = "",
-        status: str = "",
-        limit: int | None = None,
-    ) -> tuple[PrivateHealthEndpointRecord, ...]: ...
-
-
-class _RouteBindingReadStore(Protocol):
-    def read_route_binding_record(
-        self,
-        *,
-        product: str,
-        context_name: str,
-        instance_name: str,
-    ) -> EnvironmentRouteBindingRecord: ...
-
-    def list_route_binding_records(
-        self,
-        *,
-        product: str = "",
-        context_name: str = "",
-        instance_name: str = "",
-        status: str = "",
-        limit: int | None = None,
-    ) -> tuple[EnvironmentRouteBindingRecord, ...]: ...
-
-
-class _IngressCanaryRouteReadStore(Protocol):
-    def read_ingress_canary_route_record(self, canary_key: str) -> IngressCanaryRouteRecord: ...
-
-    def list_ingress_canary_route_records(
-        self,
-        *,
-        product: str = "",
-        context_name: str = "",
-        status: str = "",
-        limit: int | None = None,
-    ) -> tuple[IngressCanaryRouteRecord, ...]: ...
-
-
-class _IngressRouteAuditRecordReadStore(Protocol):
-    def read_ingress_route_audit_record(self, record_id: str) -> IngressRouteAuditRecord: ...
-
-    def list_ingress_route_audit_records(
-        self,
-        *,
-        product: str = "",
-        context_name: str = "",
-        limit: int | None = None,
-    ) -> tuple[IngressRouteAuditRecord, ...]: ...
-
-
 class _EveryCodeWorkRequestListStore(Protocol):
     def list_every_code_work_request_records(
         self,
@@ -3769,63 +3602,6 @@ def require_dokploy_target_inspect_store(record_store: object) -> DokployTargetI
     return cast(DokployTargetInspectStore, record_store)
 
 
-def require_edge_endpoint_read_store(record_store: object) -> _EdgeEndpointReadStore:
-    required_methods = (
-        "read_edge_endpoint_record",
-        "list_edge_endpoint_records",
-    )
-    missing_methods = [
-        method_name
-        for method_name in required_methods
-        if not callable(getattr(record_store, method_name, None))
-    ]
-    if missing_methods:
-        missing_summary = ", ".join(missing_methods)
-        raise TypeError(
-            f"Launchplane record store does not support edge endpoint reads: {missing_summary}"
-        )
-    return cast(_EdgeEndpointReadStore, record_store)
-
-
-def require_private_health_endpoint_read_store(
-    record_store: object,
-) -> _PrivateHealthEndpointReadStore:
-    required_methods = (
-        "read_private_health_endpoint_record",
-        "list_private_health_endpoint_records",
-    )
-    missing_methods = [
-        method_name
-        for method_name in required_methods
-        if not callable(getattr(record_store, method_name, None))
-    ]
-    if missing_methods:
-        missing_summary = ", ".join(missing_methods)
-        raise TypeError(
-            "Launchplane record store does not support private health endpoint reads: "
-            f"{missing_summary}"
-        )
-    return cast(_PrivateHealthEndpointReadStore, record_store)
-
-
-def require_route_binding_read_store(record_store: object) -> _RouteBindingReadStore:
-    required_methods = (
-        "read_route_binding_record",
-        "list_route_binding_records",
-    )
-    missing_methods = [
-        method_name
-        for method_name in required_methods
-        if not callable(getattr(record_store, method_name, None))
-    ]
-    if missing_methods:
-        missing_summary = ", ".join(missing_methods)
-        raise TypeError(
-            f"Launchplane record store does not support route binding reads: {missing_summary}"
-        )
-    return cast(_RouteBindingReadStore, record_store)
-
-
 def require_edge_endpoint_apply_store(record_store: object) -> _EdgeEndpointApplyStore:
     required_methods = (
         "write_edge_endpoint_record",
@@ -4110,77 +3886,6 @@ def require_ingress_edge_endpoint_read_store(record_store: object) -> _IngressEd
             f"Launchplane record store does not support ingress edge endpoint reads: {missing_summary}"
         )
     return cast(_IngressEdgeEndpointReadStore, record_store)
-
-
-def require_ingress_canary_route_read_store(
-    record_store: object,
-) -> _IngressCanaryRouteReadStore:
-    required_methods = (
-        "read_ingress_canary_route_record",
-        "list_ingress_canary_route_records",
-    )
-    missing_methods = [
-        method_name
-        for method_name in required_methods
-        if not callable(getattr(record_store, method_name, None))
-    ]
-    if missing_methods:
-        missing_summary = ", ".join(missing_methods)
-        raise TypeError(
-            "Launchplane record store does not support ingress canary route reads: "
-            f"{missing_summary}"
-        )
-    return cast(_IngressCanaryRouteReadStore, record_store)
-
-
-def require_ingress_route_audit_record_read_store(
-    record_store: object,
-) -> _IngressRouteAuditRecordReadStore:
-    required_methods = (
-        "read_ingress_route_audit_record",
-        "list_ingress_route_audit_records",
-    )
-    missing_methods = [
-        method_name
-        for method_name in required_methods
-        if not callable(getattr(record_store, method_name, None))
-    ]
-    if missing_methods:
-        missing_summary = ", ".join(missing_methods)
-        raise TypeError(
-            "Launchplane record store does not support ingress route audit reads: "
-            f"{missing_summary}"
-        )
-    return cast(_IngressRouteAuditRecordReadStore, record_store)
-
-
-def filter_ingress_route_audit_records(
-    records: tuple[IngressRouteAuditRecord, ...],
-    *,
-    status: str = "",
-    mode: str = "",
-    provider_host_id: int | None = None,
-    trace_id: str = "",
-    idempotency_key: str = "",
-) -> tuple[IngressRouteAuditRecord, ...]:
-    filtered_records = records
-    if status:
-        filtered_records = tuple(record for record in filtered_records if record.status == status)
-    if mode:
-        filtered_records = tuple(record for record in filtered_records if record.mode == mode)
-    if provider_host_id is not None:
-        filtered_records = tuple(
-            record for record in filtered_records if record.provider_host_id == provider_host_id
-        )
-    if trace_id:
-        filtered_records = tuple(
-            record for record in filtered_records if record.trace_id == trace_id
-        )
-    if idempotency_key:
-        filtered_records = tuple(
-            record for record in filtered_records if record.idempotency_key == idempotency_key
-        )
-    return filtered_records
 
 
 def product_profile_context_cutover_allowed_contexts(
@@ -11779,246 +11484,6 @@ def create_launchplane_fastapi_app(
             {"status": "ok", "trace_id": trace_id, **logs_payload}
         )
 
-    def ensure_edge_endpoint_read_allowed(
-        *,
-        identity: LaunchplaneIdentity,
-        trace_id: str,
-    ) -> None:
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action="edge_endpoint.read",
-            product="launchplane",
-            context=_LAUNCHPLANE_SERVICE_CONTEXT,
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message="Workflow cannot read Launchplane edge endpoint records.",
-            )
-
-    def list_edge_endpoint_records(
-        identity: Annotated[LaunchplaneIdentity, Depends(read_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        limit: Annotated[str, Query()] = "25",
-        provider: Annotated[str, Query()] = "",
-        status: Annotated[str, Query()] = "",
-    ) -> EdgeEndpointRecordsResponse:
-        trace_id = next_trace_id()
-        ensure_edge_endpoint_read_allowed(identity=identity, trace_id=trace_id)
-        try:
-            normalized_limit = control_plane_service_status.query_int_value(
-                limit,
-                "limit",
-                default=25,
-                minimum=1,
-                maximum=100,
-            )
-            assert normalized_limit is not None
-        except ValueError as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_query",
-                message=str(error),
-            ) from error
-        try:
-            edge_endpoint_store = require_edge_endpoint_read_store(record_store)
-            records = edge_endpoint_store.list_edge_endpoint_records(
-                provider=provider.strip(),
-                status=status.strip(),
-                limit=normalized_limit,
-            )
-        except TypeError as error:
-            raise _launchplane_http_error(
-                status_code=503,
-                trace_id=trace_id,
-                code="database_storage_required",
-                message=str(error),
-            ) from error
-        return EdgeEndpointRecordsResponse(
-            trace_id=trace_id,
-            limit=normalized_limit,
-            count=len(records),
-            records=records,
-        )
-
-    def read_edge_endpoint_record(
-        endpoint_key: Annotated[str, Path(min_length=1, pattern=r"^\S+$")],
-        identity: Annotated[LaunchplaneIdentity, Depends(read_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-    ) -> EdgeEndpointRecordResponse:
-        trace_id = next_trace_id()
-        ensure_edge_endpoint_read_allowed(identity=identity, trace_id=trace_id)
-        try:
-            edge_endpoint_store = require_edge_endpoint_read_store(record_store)
-            record = edge_endpoint_store.read_edge_endpoint_record(endpoint_key)
-        except TypeError as error:
-            raise _launchplane_http_error(
-                status_code=503,
-                trace_id=trace_id,
-                code="database_storage_required",
-                message=str(error),
-            ) from error
-        except FileNotFoundError as error:
-            raise _launchplane_http_error(
-                status_code=404,
-                trace_id=trace_id,
-                code="not_found",
-                message=str(error),
-            ) from error
-        return EdgeEndpointRecordResponse(trace_id=trace_id, record=record)
-
-    def ensure_private_health_endpoint_read_allowed(
-        *,
-        identity: LaunchplaneIdentity,
-        trace_id: str,
-        product: str,
-        context_name: str,
-    ) -> None:
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action="private_health_endpoint.read",
-            product=product,
-            context=context_name,
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message=(
-                    "Workflow cannot read private health endpoints for the requested "
-                    "product/context."
-                ),
-            )
-
-    def list_private_health_endpoint_records(
-        identity: Annotated[LaunchplaneIdentity, Depends(read_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        product: Annotated[str, Query()] = "",
-        context: Annotated[str, Query()] = "",
-        instance: Annotated[str, Query()] = "",
-        status: Annotated[str, Query()] = "",
-        limit: Annotated[str, Query()] = "25",
-    ) -> PrivateHealthEndpointRecordsResponse:
-        trace_id = next_trace_id()
-        normalized_product = product.strip()
-        context_name = context.strip()
-        instance_name = instance.strip()
-        if not normalized_product or not context_name:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_query",
-                message=(
-                    "Private health endpoint reads require product and context query parameters."
-                ),
-            )
-        ensure_private_health_endpoint_read_allowed(
-            identity=identity,
-            trace_id=trace_id,
-            product=normalized_product,
-            context_name=context_name,
-        )
-        try:
-            normalized_limit = control_plane_service_status.query_int_value(
-                limit,
-                "limit",
-                default=25,
-                minimum=1,
-                maximum=100,
-            )
-            assert normalized_limit is not None
-        except ValueError as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_query",
-                message=str(error),
-            ) from error
-        try:
-            private_endpoint_store = require_private_health_endpoint_read_store(record_store)
-            records = private_endpoint_store.list_private_health_endpoint_records(
-                product=normalized_product,
-                context_name=context_name,
-                instance_name=instance_name,
-                status=status.strip(),
-                limit=normalized_limit,
-            )
-        except TypeError as error:
-            raise _launchplane_http_error(
-                status_code=503,
-                trace_id=trace_id,
-                code="database_storage_required",
-                message=str(error),
-            ) from error
-        return PrivateHealthEndpointRecordsResponse(
-            trace_id=trace_id,
-            product=normalized_product,
-            context=context_name,
-            instance=instance_name,
-            limit=normalized_limit,
-            count=len(records),
-            records=records,
-        )
-
-    def read_private_health_endpoint_record(
-        endpoint_key: Annotated[str, Path(min_length=1, pattern=r"^\S+$")],
-        identity: Annotated[LaunchplaneIdentity, Depends(read_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        product: Annotated[str, Query()] = "",
-        context: Annotated[str, Query()] = "",
-        instance: Annotated[str, Query()] = "",
-    ) -> PrivateHealthEndpointRecordResponse:
-        trace_id = next_trace_id()
-        normalized_product = product.strip()
-        context_name = context.strip()
-        instance_name = instance.strip()
-        if not normalized_product or not context_name:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_query",
-                message=(
-                    "Private health endpoint reads require product and context query parameters."
-                ),
-            )
-        ensure_private_health_endpoint_read_allowed(
-            identity=identity,
-            trace_id=trace_id,
-            product=normalized_product,
-            context_name=context_name,
-        )
-        try:
-            private_endpoint_store = require_private_health_endpoint_read_store(record_store)
-            record = private_endpoint_store.read_private_health_endpoint_record(endpoint_key)
-        except TypeError as error:
-            raise _launchplane_http_error(
-                status_code=503,
-                trace_id=trace_id,
-                code="database_storage_required",
-                message=str(error),
-            ) from error
-        except FileNotFoundError as error:
-            raise _launchplane_http_error(
-                status_code=404,
-                trace_id=trace_id,
-                code="not_found",
-                message=str(error),
-            ) from error
-        if (
-            record.product != normalized_product
-            or record.context != context_name
-            or (instance_name and record.instance != instance_name)
-        ):
-            raise _launchplane_http_error(
-                status_code=404,
-                trace_id=trace_id,
-                code="not_found",
-                message=f"Record not found: {endpoint_key}",
-            )
-        return PrivateHealthEndpointRecordResponse(trace_id=trace_id, record=record)
-
     def ensure_route_binding_allowed(
         *,
         identity: LaunchplaneIdentity,
@@ -12040,372 +11505,6 @@ def create_launchplane_fastapi_app(
                 code="authorization_denied",
                 message=message,
             )
-
-    def list_route_binding_records(
-        identity: Annotated[LaunchplaneIdentity, Depends(read_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        product: Annotated[str, Query()] = "",
-        context: Annotated[str, Query()] = "",
-        instance: Annotated[str, Query()] = "",
-        status: Annotated[str, Query()] = "",
-        limit: Annotated[str, Query()] = "25",
-    ) -> RouteBindingRecordsResponse:
-        trace_id = next_trace_id()
-        normalized_product = product.strip()
-        context_name = context.strip()
-        instance_name = instance.strip()
-        if not normalized_product or not context_name:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_query",
-                message="Route binding list requires product and context query parameters.",
-            )
-        ensure_route_binding_allowed(
-            identity=identity,
-            trace_id=trace_id,
-            action="route_binding.read",
-            product=normalized_product,
-            context_name=context_name,
-            message="Workflow cannot read route bindings for the requested product/context.",
-        )
-        try:
-            normalized_limit = control_plane_service_status.query_int_value(
-                limit,
-                "limit",
-                default=25,
-                minimum=1,
-                maximum=100,
-            )
-            assert normalized_limit is not None
-            route_binding_store = require_route_binding_read_store(record_store)
-            records = route_binding_store.list_route_binding_records(
-                product=normalized_product,
-                context_name=context_name,
-                instance_name=instance_name,
-                status=status.strip(),
-                limit=normalized_limit,
-            )
-        except ValueError as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_query",
-                message=str(error),
-            ) from error
-        except TypeError as error:
-            raise _launchplane_http_error(
-                status_code=503,
-                trace_id=trace_id,
-                code="database_storage_required",
-                message=str(error),
-            ) from error
-        return RouteBindingRecordsResponse(
-            trace_id=trace_id,
-            product=normalized_product,
-            context=context_name,
-            instance=instance_name,
-            limit=normalized_limit,
-            count=len(records),
-            records=tuple(redacted_route_binding_record(record) for record in records),
-        )
-
-    def read_route_binding_record(
-        identity: Annotated[LaunchplaneIdentity, Depends(read_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        product: Annotated[str, Query()] = "",
-        context: Annotated[str, Query()] = "",
-        instance: Annotated[str, Query()] = "",
-    ) -> RouteBindingRecordResponse:
-        trace_id = next_trace_id()
-        normalized_product = product.strip()
-        context_name = context.strip()
-        instance_name = instance.strip()
-        if not normalized_product or not context_name or not instance_name:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_query",
-                message=(
-                    "Route binding record reads require product, context, and instance "
-                    "query parameters."
-                ),
-            )
-        ensure_route_binding_allowed(
-            identity=identity,
-            trace_id=trace_id,
-            action="route_binding.read",
-            product=normalized_product,
-            context_name=context_name,
-            message="Workflow cannot read route bindings for the requested product/context.",
-        )
-        try:
-            route_binding_store = require_route_binding_read_store(record_store)
-            record = route_binding_store.read_route_binding_record(
-                product=normalized_product,
-                context_name=context_name,
-                instance_name=instance_name,
-            )
-        except TypeError as error:
-            raise _launchplane_http_error(
-                status_code=503,
-                trace_id=trace_id,
-                code="database_storage_required",
-                message=str(error),
-            ) from error
-        except FileNotFoundError as error:
-            raise _launchplane_http_error(
-                status_code=404,
-                trace_id=trace_id,
-                code="not_found",
-                message=str(error),
-            ) from error
-        return RouteBindingRecordResponse(
-            trace_id=trace_id,
-            record=redacted_route_binding_record(record),
-        )
-
-    def ensure_ingress_canary_route_read_allowed(
-        *,
-        identity: LaunchplaneIdentity,
-        trace_id: str,
-    ) -> None:
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action="ingress_canary_route.read",
-            product="launchplane",
-            context=_LAUNCHPLANE_SERVICE_CONTEXT,
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message="Workflow cannot read Launchplane ingress canary route records.",
-            )
-
-    def list_ingress_canary_route_records(
-        identity: Annotated[LaunchplaneIdentity, Depends(read_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        limit: Annotated[str, Query()] = "25",
-        product: Annotated[str, Query()] = "",
-        context: Annotated[str, Query()] = "",
-        status: Annotated[str, Query()] = "",
-    ) -> IngressCanaryRouteRecordsResponse:
-        trace_id = next_trace_id()
-        ensure_ingress_canary_route_read_allowed(identity=identity, trace_id=trace_id)
-        try:
-            normalized_limit = control_plane_service_status.query_int_value(
-                limit,
-                "limit",
-                default=25,
-                minimum=1,
-                maximum=100,
-            )
-            assert normalized_limit is not None
-        except ValueError as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_query",
-                message=str(error),
-            ) from error
-        try:
-            canary_store = require_ingress_canary_route_read_store(record_store)
-            records = canary_store.list_ingress_canary_route_records(
-                product=product.strip(),
-                context_name=context.strip(),
-                status=status.strip(),
-                limit=normalized_limit,
-            )
-        except TypeError as error:
-            raise _launchplane_http_error(
-                status_code=503,
-                trace_id=trace_id,
-                code="database_storage_required",
-                message=str(error),
-            ) from error
-        return IngressCanaryRouteRecordsResponse(
-            trace_id=trace_id,
-            limit=normalized_limit,
-            count=len(records),
-            records=records,
-        )
-
-    def read_ingress_canary_route_record(
-        canary_key: Annotated[str, Path(min_length=1, pattern=r"^\S+$")],
-        identity: Annotated[LaunchplaneIdentity, Depends(read_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-    ) -> IngressCanaryRouteRecordResponse:
-        trace_id = next_trace_id()
-        ensure_ingress_canary_route_read_allowed(identity=identity, trace_id=trace_id)
-        try:
-            canary_store = require_ingress_canary_route_read_store(record_store)
-            record = canary_store.read_ingress_canary_route_record(canary_key)
-        except TypeError as error:
-            raise _launchplane_http_error(
-                status_code=503,
-                trace_id=trace_id,
-                code="database_storage_required",
-                message=str(error),
-            ) from error
-        except FileNotFoundError as error:
-            raise _launchplane_http_error(
-                status_code=404,
-                trace_id=trace_id,
-                code="not_found",
-                message=str(error),
-            ) from error
-        return IngressCanaryRouteRecordResponse(trace_id=trace_id, record=record)
-
-    def ensure_ingress_route_audit_read_allowed(
-        *,
-        identity: LaunchplaneIdentity,
-        trace_id: str,
-        product: str,
-        context_name: str,
-    ) -> None:
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action="ingress_route.plan",
-            product=product,
-            context=context_name,
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message="Workflow cannot read ingress route audit records for the requested product/context.",
-            )
-
-    def list_ingress_route_audit_records(
-        identity: Annotated[LaunchplaneIdentity, Depends(read_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        product: Annotated[str, Query()] = "",
-        context: Annotated[str, Query()] = "",
-        status: Annotated[str, Query()] = "",
-        mode: Annotated[str, Query()] = "",
-        provider_host_id: Annotated[str, Query()] = "",
-        trace_id: Annotated[str, Query()] = "",
-        idempotency_key: Annotated[str, Query()] = "",
-        limit: Annotated[str, Query()] = "25",
-    ) -> IngressRouteAuditRecordsResponse:
-        request_trace_id = next_trace_id()
-        normalized_product = product.strip()
-        context_name = context.strip()
-        if not normalized_product or not context_name:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=request_trace_id,
-                code="invalid_query",
-                message="Ingress route audit list requires product and context query parameters.",
-            )
-        ensure_ingress_route_audit_read_allowed(
-            identity=identity,
-            trace_id=request_trace_id,
-            product=normalized_product,
-            context_name=context_name,
-        )
-        try:
-            normalized_limit = control_plane_service_status.query_int_value(
-                limit,
-                "limit",
-                default=25,
-                minimum=1,
-                maximum=100,
-            )
-            assert normalized_limit is not None
-            normalized_provider_host_id = control_plane_service_status.query_int_value(
-                provider_host_id,
-                "provider_host_id",
-                minimum=1,
-            )
-        except ValueError as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=request_trace_id,
-                code="invalid_query",
-                message=str(error),
-            ) from error
-        try:
-            audit_store = require_ingress_route_audit_record_read_store(record_store)
-            records = audit_store.list_ingress_route_audit_records(
-                product=normalized_product,
-                context_name=context_name,
-                limit=None,
-            )
-        except TypeError as error:
-            raise _launchplane_http_error(
-                status_code=503,
-                trace_id=request_trace_id,
-                code="database_storage_required",
-                message=str(error),
-            ) from error
-        records = filter_ingress_route_audit_records(
-            records,
-            status=status.strip(),
-            mode=mode.strip(),
-            provider_host_id=normalized_provider_host_id,
-            trace_id=trace_id.strip(),
-            idempotency_key=idempotency_key.strip(),
-        )
-        limited_records = records[:normalized_limit]
-        return IngressRouteAuditRecordsResponse(
-            trace_id=request_trace_id,
-            product=normalized_product,
-            context=context_name,
-            limit=normalized_limit,
-            count=len(limited_records),
-            records=limited_records,
-        )
-
-    def read_ingress_route_audit_record(
-        record_id: Annotated[str, Path(min_length=1, pattern=r"^\S+$")],
-        identity: Annotated[LaunchplaneIdentity, Depends(read_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        product: Annotated[str, Query()] = "",
-        context: Annotated[str, Query()] = "",
-    ) -> IngressRouteAuditRecordResponse:
-        request_trace_id = next_trace_id()
-        normalized_product = product.strip()
-        context_name = context.strip()
-        if not normalized_product or not context_name:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=request_trace_id,
-                code="invalid_query",
-                message="Ingress route audit record reads require product and context query parameters.",
-            )
-        ensure_ingress_route_audit_read_allowed(
-            identity=identity,
-            trace_id=request_trace_id,
-            product=normalized_product,
-            context_name=context_name,
-        )
-        try:
-            audit_store = require_ingress_route_audit_record_read_store(record_store)
-            record = audit_store.read_ingress_route_audit_record(record_id)
-        except TypeError as error:
-            raise _launchplane_http_error(
-                status_code=503,
-                trace_id=request_trace_id,
-                code="database_storage_required",
-                message=str(error),
-            ) from error
-        except FileNotFoundError as error:
-            raise _launchplane_http_error(
-                status_code=404,
-                trace_id=request_trace_id,
-                code="not_found",
-                message=str(error),
-            ) from error
-        if record.product != normalized_product or record.context != context_name:
-            raise _launchplane_http_error(
-                status_code=404,
-                trace_id=request_trace_id,
-                code="not_found",
-                message=f"Record not found: {record_id}",
-            )
-        return IngressRouteAuditRecordResponse(trace_id=request_trace_id, record=record)
 
     def read_deployment_record(
         record_id: Annotated[str, Path(min_length=1, pattern=r"^\S+$")],
@@ -21680,158 +20779,30 @@ def create_launchplane_fastapi_app(
         },
     )
 
-    app.add_api_route(
-        "/v1/route-bindings/records",
-        list_route_binding_records,
-        methods=["GET"],
-        response_model=RouteBindingRecordsResponse,
-        operation_id="list_route_binding_records",
-        summary="List provider-neutral environment route bindings",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
+    def read_route_authorization_allows(
+        *,
+        identity: LaunchplaneIdentity,
+        action: str,
+        product: str,
+        context: str,
+    ) -> bool:
+        return resolved_authz_policy_runtime.policy.allows(
+            identity=identity,
+            action=action,
+            product=product,
+            context=context,
+        )
 
-    app.add_api_route(
-        "/v1/route-bindings/records/current",
-        read_route_binding_record,
-        methods=["GET"],
-        response_model=RouteBindingRecordResponse,
-        operation_id="read_route_binding_record",
-        summary="Read one provider-neutral environment route binding",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            404: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
+    read_route_dependencies = ReadRouteDependencies(
+        read_identity=read_identity,
+        get_record_store=get_record_store,
+        next_trace_id=next_trace_id,
+        authorization_allows=read_route_authorization_allows,
+        http_error=_launchplane_http_error,
+        error_response_model=LaunchplaneErrorResponse,
     )
-
-    app.add_api_route(
-        "/v1/edge-endpoints/records",
-        list_edge_endpoint_records,
-        methods=["GET"],
-        response_model=EdgeEndpointRecordsResponse,
-        operation_id="list_edge_endpoint_records",
-        summary="List edge endpoint records",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        "/v1/edge-endpoints/records/{endpoint_key}",
-        read_edge_endpoint_record,
-        methods=["GET"],
-        response_model=EdgeEndpointRecordResponse,
-        operation_id="read_edge_endpoint_record",
-        summary="Read one edge endpoint record",
-        responses={
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            404: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        "/v1/private-health-endpoints/records",
-        list_private_health_endpoint_records,
-        methods=["GET"],
-        response_model=PrivateHealthEndpointRecordsResponse,
-        operation_id="list_private_health_endpoint_records",
-        summary="List private health endpoint records",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        "/v1/private-health-endpoints/records/{endpoint_key}",
-        read_private_health_endpoint_record,
-        methods=["GET"],
-        response_model=PrivateHealthEndpointRecordResponse,
-        operation_id="read_private_health_endpoint_record",
-        summary="Read one private health endpoint record",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            404: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        "/v1/ingress/canary-routes/records",
-        list_ingress_canary_route_records,
-        methods=["GET"],
-        response_model=IngressCanaryRouteRecordsResponse,
-        operation_id="list_ingress_canary_route_records",
-        summary="List ingress canary route records",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        "/v1/ingress/canary-routes/records/{canary_key}",
-        read_ingress_canary_route_record,
-        methods=["GET"],
-        response_model=IngressCanaryRouteRecordResponse,
-        operation_id="read_ingress_canary_route_record",
-        summary="Read one ingress canary route record",
-        responses={
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            404: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        "/v1/ingress/route-audits/records",
-        list_ingress_route_audit_records,
-        methods=["GET"],
-        response_model=IngressRouteAuditRecordsResponse,
-        operation_id="list_ingress_route_audit_records",
-        summary="List ingress route audit records",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        "/v1/ingress/route-audits/records/{record_id}",
-        read_ingress_route_audit_record,
-        methods=["GET"],
-        response_model=IngressRouteAuditRecordResponse,
-        operation_id="read_ingress_route_audit_record",
-        summary="Read one ingress route audit record",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            404: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
+    register_topology_read_routes(app, dependencies=read_route_dependencies)
+    register_ingress_read_routes(app, dependencies=read_route_dependencies)
 
     app.add_api_route(
         "/v1/deployments/{record_id}",
