+++ control_plane/http_routes/__init__.py
@@ -1,0 +1,9 @@
+from control_plane.http_routes.ingress import register_ingress_read_routes
+from control_plane.http_routes.support import ReadRouteDependencies
+from control_plane.http_routes.topology import register_topology_read_routes
+
+__all__ = (
+    "ReadRouteDependencies",
+    "register_ingress_read_routes",
+    "register_topology_read_routes",
+)
