+++ tests/http_app_test_support.py
@@ -36,10 +36,6 @@ from control_plane.contracts.every_code_work_request import (
     EveryCodeWorkRequestRecord,
 )
 from control_plane.contracts.idempotency_record import LaunchplaneIdempotencyRecord
-from control_plane.contracts.ingress_route_audit_record import (
-    IngressRouteAuditOperation,
-    IngressRouteAuditRecord,
-)
 from control_plane.contracts.merge_train_policy import (
     MergeTrainPolicyRecord,
 )
@@ -1301,25 +1297,6 @@ def _record_read_policy(
     )
 
 
-def _private_health_endpoint_read_policy() -> LaunchplaneAuthzPolicy:
-    return LaunchplaneAuthzPolicy.model_validate(
-        {
-            "github_actions": [
-                {
-                    "repository": "every/verireel",
-                    "workflow_refs": [
-                        "every/verireel/.github/workflows/preview-control-plane.yml@refs/heads/main"
-                    ],
-                    "event_names": ["pull_request"],
-                    "products": ["repairshopr-sync"],
-                    "contexts": ["repairshopr-sync"],
-                    "actions": ["private_health_endpoint.read"],
-                }
-            ]
-        }
-    )
-
-
 def _github_human_deployment_write_policy(*, context: str) -> LaunchplaneAuthzPolicy:
     return LaunchplaneAuthzPolicy.model_validate(
         {
@@ -2332,65 +2309,6 @@ def _preview_pr_feedback_payload(
     if dry_run:
         payload["dry_run"] = True
     return payload
-
-
-def _ingress_route_audit_record(
-    *,
-    record_id: str = "ingress-route-audit-test",
-    product: str = "launchplane",
-    context: str = "reon-prod",
-    mode: Literal["dry-run", "apply"] = "dry-run",
-    status: Literal["pending", "planned", "applied", "unchanged"] = "planned",
-    dry_run: bool = True,
-    provider_host_id: int | None = 78,
-    trace_id: str = "trace-audit-1",
-    idempotency_key: str = "audit-key-1",
-    recorded_at: str = "2026-06-01T00:00:00Z",
-) -> IngressRouteAuditRecord:
-    return IngressRouteAuditRecord(
-        record_id=record_id,
-        product=product,
-        context=context,
-        mode=mode,
-        status=status,
-        dry_run=dry_run,
-        requested_domains=("app.example.com",),
-        edge_endpoint_key="edge-app",
-        expected_host_id=None,
-        provider_host_id=provider_host_id,
-        operations=(
-            IngressRouteAuditOperation(
-                action="create",
-                host_id=provider_host_id,
-                domain_names=("app.example.com",),
-                requires_apply=mode == "dry-run",
-                change_categories=("create",),
-            ),
-        ),
-        trace_id=trace_id,
-        idempotency_key=idempotency_key,
-        reason="test",
-        recorded_at=recorded_at,
-    )
-
-
-def _ingress_route_audit_read_policy(*, contexts: tuple[str, ...]) -> LaunchplaneAuthzPolicy:
-    return LaunchplaneAuthzPolicy.model_validate(
-        {
-            "github_actions": [
-                {
-                    "repository": "every/verireel",
-                    "workflow_refs": [
-                        "every/verireel/.github/workflows/preview-control-plane.yml@refs/heads/main"
-                    ],
-                    "event_names": ["pull_request"],
-                    "products": ["launchplane"],
-                    "contexts": list(contexts),
-                    "actions": ["ingress_route.plan"],
-                }
-            ]
-        }
-    )
 
 
 def _product_profile_read_policy(*, product: str) -> LaunchplaneAuthzPolicy:
@@ -4023,218 +3941,6 @@ async def _get_tracked_target_logs(
     )
 
 
-async def _get_edge_endpoint_records(
-    app: FastAPI,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-    limit: str = "",
-    provider: str = "",
-    status: str = "",
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    params = {}
-    if limit:
-        params["limit"] = limit
-    if provider:
-        params["provider"] = provider
-    if status:
-        params["status"] = status
-    suffix = f"?{urlencode(params)}" if params else ""
-    return await _asgi_get(
-        app,
-        f"/v1/edge-endpoints/records{suffix}",
-        headers=request_headers,
-    )
-
-
-async def _get_edge_endpoint_record(
-    app: FastAPI,
-    endpoint_key: str,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    return await _asgi_get(
-        app,
-        f"/v1/edge-endpoints/records/{endpoint_key}",
-        headers=request_headers,
-    )
-
-
-async def _get_private_health_endpoint_records(
-    app: FastAPI,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-    product: str = "repairshopr-sync",
-    context: str = "repairshopr-sync",
-    instance: str = "",
-    status: str = "",
-    limit: str = "",
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    params = {}
-    if product:
-        params["product"] = product
-    if context:
-        params["context"] = context
-    if instance:
-        params["instance"] = instance
-    if status:
-        params["status"] = status
-    if limit:
-        params["limit"] = limit
-    suffix = f"?{urlencode(params)}" if params else ""
-    return await _asgi_get(
-        app,
-        f"/v1/private-health-endpoints/records{suffix}",
-        headers=request_headers,
-    )
-
-
-async def _get_private_health_endpoint_record(
-    app: FastAPI,
-    endpoint_key: str,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-    product: str = "repairshopr-sync",
-    context: str = "repairshopr-sync",
-    instance: str = "prod",
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    params = {}
-    if product:
-        params["product"] = product
-    if context:
-        params["context"] = context
-    if instance:
-        params["instance"] = instance
-    suffix = f"?{urlencode(params)}" if params else ""
-    return await _asgi_get(
-        app,
-        f"/v1/private-health-endpoints/records/{endpoint_key}{suffix}",
-        headers=request_headers,
-    )
-
-
-async def _get_route_binding_records(
-    app: FastAPI,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-    product: str = "example-product",
-    context: str = "example-testing",
-    instance: str = "",
-    status: str = "",
-    limit: str = "",
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    params = {}
-    if product:
-        params["product"] = product
-    if context:
-        params["context"] = context
-    if instance:
-        params["instance"] = instance
-    if status:
-        params["status"] = status
-    if limit:
-        params["limit"] = limit
-    suffix = f"?{urlencode(params)}" if params else ""
-    return await _asgi_get(
-        app,
-        f"/v1/route-bindings/records{suffix}",
-        headers=request_headers,
-    )
-
-
-async def _get_route_binding_record(
-    app: FastAPI,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-    product: str = "example-product",
-    context: str = "example-testing",
-    instance: str = "web",
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    params = {}
-    if product:
-        params["product"] = product
-    if context:
-        params["context"] = context
-    if instance:
-        params["instance"] = instance
-    suffix = f"?{urlencode(params)}" if params else ""
-    return await _asgi_get(
-        app,
-        f"/v1/route-bindings/records/current{suffix}",
-        headers=request_headers,
-    )
-
-
-async def _get_ingress_canary_route_records(
-    app: FastAPI,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-    product: str = "",
-    context: str = "",
-    status: str = "",
-    limit: str = "",
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    params = {}
-    if product:
-        params["product"] = product
-    if context:
-        params["context"] = context
-    if status:
-        params["status"] = status
-    if limit:
-        params["limit"] = limit
-    suffix = f"?{urlencode(params)}" if params else ""
-    return await _asgi_get(
-        app,
-        f"/v1/ingress/canary-routes/records{suffix}",
-        headers=request_headers,
-    )
-
-
-async def _get_ingress_canary_route_record(
-    app: FastAPI,
-    canary_key: str,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    return await _asgi_get(
-        app,
-        f"/v1/ingress/canary-routes/records/{canary_key}",
-        headers=request_headers,
-    )
-
-
 def _ingress_canary_route_apply_payload(
     *,
     product: str = "launchplane",
@@ -4249,62 +3955,6 @@ def _ingress_canary_route_apply_payload(
         "canary_key": canary_key,
         "reason": reason,
     }
-
-
-async def _get_ingress_route_audit_records(
-    app: FastAPI,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-    product: str = "",
-    context: str = "",
-    status: str = "",
-    mode: str = "",
-    provider_host_id: str = "",
-    trace_id: str = "",
-    idempotency_key: str = "",
-    limit: str = "",
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    params = _query_params(
-        product=product,
-        context=context,
-        status=status,
-        mode=mode,
-        provider_host_id=provider_host_id,
-        trace_id=trace_id,
-        idempotency_key=idempotency_key,
-        limit=limit,
-    )
-    suffix = f"?{urlencode(params)}" if params else ""
-    return await _asgi_get(
-        app,
-        f"/v1/ingress/route-audits/records{suffix}",
-        headers=request_headers,
-    )
-
-
-async def _get_ingress_route_audit_record(
-    app: FastAPI,
-    record_id: str,
-    *,
-    authorization: str = "Bearer valid-token",
-    headers: dict[str, str] | None = None,
-    product: str = "",
-    context: str = "",
-) -> _AsgiResponse:
-    request_headers = dict(headers or {})
-    if authorization:
-        request_headers["Authorization"] = authorization
-    params = _query_params(product=product, context=context)
-    suffix = f"?{urlencode(params)}" if params else ""
-    return await _asgi_get(
-        app,
-        f"/v1/ingress/route-audits/records/{record_id}{suffix}",
-        headers=request_headers,
-    )
 
 
 async def _get_deployment_record(
